Add option:  sgrid --argsfile args.txt
If we start sgrid like this, sgrid opens the file args.txt and uses
the lines in there as regular sgrid args (parfile and options).
If there are several lines sgrid will restart itself. These restarts
are fragile: All static C-vars are unchanged on restart! Vars and Funs
are also unchanged! This is wrong if certain Pars change on restart!
But it will work fine if all that changes between restarts are just
some pars that do not affect the grid or the number of Vars and such.