Python script to fix output bug in old bam waves
I am pretty sure now that bitant symmetry was incorrectly applied in
shell box5 in bam, for the region where the 3rd shell coordinate is
negative. This results in errors in the region 0<theta<pi/4, phi>pi in
the sphere output when using bitant in bam, because it took the data
from region where the 3rd shell coordinate is positive and maps it into
the region where it is negative.

Since we have lots of data that we would like to fix, we can just
figure out what symmetry rpsi4*.l0 and ipsi4*.l0 should have and fill in
the region 0<theta<pi/4, phi>pi using that symmetry. After this is done,
we can run sgrid and get all the modes. This assumes that they have
some symmetry across the zero line of the 3rd shell coord in box5.

In quadrantSymFix.py I have simply assumed that both rpsi4*.l0 and ipsi4*.l0
have quadrant symmetry. That allows us to fill in the region with the errors.
This should work as long as we have equal stars that stay in the initial
orbital plane.