change things like "int touch : 1;" to "unsigned touch : 1;" in tBface
This is more logical since then the possible values of touch are 0
and 1. When we use int the possible values are 0 and -1, because a
signed int with only one bit cannot store +1. (Although I guess the
actual bit pattern in memory is probably the same when the machine
uses 2's complement for negative numbers.)