add same_fpts bit to bface, add tCoordInfo CI[1] in each box
same_fpts=1 if the ith point in fpts is at the same (x,y,z) as the ith point
in the fpts of the corresponding bface of the other box. I.e. there are the
same number of points in both bfaces and all their locations agree. This
simplities interbox BCs.

The tCoordInfo CI contains additional (optional) info about the coords of each
box. This info can be used in the coord. transform.