add more printf in case the bface setup goes wrong
If one bface has touch=1 and the other has touch=0 we errorexit.

If same_fpts=0 and touch=0 in one bface, at least one point in the pointlist
is not on the face of the other box. For Coordinates_set_bfaces_oldWT this
can be caused by set_bfaces_on_boxface not really checking if edges or
corners really belong to the bface. Then set_bits_in_all_bfaces cannot find
these points on the face of the other box - see the lines with:

          if(face[of]) N++;
and

      if(fpts->npoints[box->b] == N)

in func: int set_bits_in_all_bfaces(tGrid *grid)